add lavamoat runtime for packaged apps
Currently electron-builder messes around
with the package.json as well as the node_modules
(for example it does not consider the dev deps
node_modules to be copied to the packaged app).
Due to that we have generated a policy.json
only for packaged electron apps.
Also it seems that lavamoat computes different
paths when running within electron. Due to that
we are adding a specific policy override file.
Finally we had to patch lavamoat to consider
leveldown as a builtin when the app is packaged only.