feat: add desktop ui build scripts
The scripts are pretty much an adaptation
from the Extension build scripts. There is no
point in reinventing the wheel, especially when we
want to have the same lavamoat protections (and due
to that we need to use browserify to produce the bundle).
There are some stuff that we need to clean up on
subsequent PRs, like having our own Desktop
constants and tasks, have our own static assets
(locales, fonts, css, etc..)