chore: add .env for desktop e2e tests (#282)
We need a .env for e2e tests
because whenever we want to change the
electron app entry path, we need to
change the ELECTRON_APP_PATH env var
for the e2e tests. If that var only exists
on the CI, then updating it will:

- allow the branch that changes the
entry path to pass

- fail all other branches

For that reason we need to specify
the entry path on the branch itself.