chore: move desktop ui from extension
We are moving desktop ui outside
of the extension. This is within
the effort of moving the extension
codebase to a git submodule.
we also had to move all the ui build
scripts and do some minor changes like:
* add babel-ui config for browserify bundle
* remove browserify code fencing as it was
using linting from the extension. And start
using the same code fencing as we are using
on the main process build.
* Add a index.scss to be the entry point
for desktop ui scss (it imports the extension
main scss file)
* update the build styles task to include both
ui and extension ui paths
* add existing ui lint rules from extension to
eslintrc