fix: copy Dropdown component from extension
While using extension components
there is a risk of the UI referencing
two different versions of react
(https://reactjs.org/warnings/invalid-hook-call-warning.html\#duplicate-react)
This cannot be solved as the extension
submodule is not a library being imported.
So adding resolutions to package.json
would not solve the problem. The only way
is by stop referencing the extension
components and have our own Desktop
components. We are moving DropDown
as that one was explicitly breaking,
but this should be just the first one
that we move out.