fix: rename babel.config used for app
Currently babelify used on the ui
bundle scripts, does not allow
specify a path for a config file.
It just allows to specify an config
file to extend the base one. This means
that the current babel config for ui
would be a merge between babel.config.js
+ babel-ui.config.js. As that is something
that we do not want, we are renaming
babel.config.js to babel-app.config.js.
This way the app one is used for the app build
and the ui one is used for the ui build.