Move desktop UI outside of extension (#296)
* chore: move desktop ui from extension

We are moving desktop ui outside
of the extension. This is within
the effort of moving the extension
codebase to a git submodule.
we also had to move all the ui build
scripts and do some minor changes like:
* add babel-ui config for browserify bundle
* remove browserify code fencing as it was
using linting from the extension. And start
using the same code fencing as we are using
on the main process build.
* Add a index.scss to be the entry point
for desktop ui scss (it imports the extension
main scss file)
* update the build styles task to include both
ui and extension ui paths
* add existing ui lint rules from extension to
eslintrc

* fix: copy Dropdown component from extension

While using extension components
there is a risk of the UI referencing
two different versions of react
(https://reactjs.org/warnings/invalid-hook-call-warning.html\#duplicate-react)
This cannot be solved as the extension
submodule is not a library being imported.
So adding resolutions to package.json
would not solve the problem. The only way
is by stop referencing the extension
components and have our own Desktop
components. We are moving DropDown
as that one was explicitly breaking,
but this should be just the first one
that we move out.

* chore: lavammoat aa allow extension node modules

Patching lavamoat aa to allow
referencing extension node modules.
This happens because lavamoat browserify
resolves require's inside the extension code
to the extension node_modules.

* fix: remove unused deps from extension

* fix: rename babel.config used for app

Currently babelify used on the ui
bundle scripts, does not allow
specify a path for a config file.
It just allows to specify an config
file to extend the base one. This means
that the current babel config for ui
would be a merge between babel.config.js
+ babel-ui.config.js. As that is something
that we do not want, we are renaming
babel.config.js to babel-app.config.js.
This way the app one is used for the app build
and the ui one is used for the ui build.

* fix: add babel config for jest tests

* chore: update eslint config composition

Updated it in order to have a more
modular approach, where we can define
rules per file type for src, ui and test
directories. The exception are tests
living side by side with the code.