chore: add tag name option to create release branch cli
Add tag name option that will
allow to not only create the release
branch, but also tag that commit id
from where the release branch is created.
We need this for the app, in order to simplify
creating the release branch for the desktop app
and automatically set a correct changelog.
This is important as we do not want to create
confusion between actual app releases, and
preparation for releases (i.e. cut off from dev
and start QA)