chore: allow manually bypassing is release
We want to do this for testing purposes.
Testing the package app locally with act
is not possible (act is super limited.. for
example it does not support multiple OS, only linux).
For that reason we want to test the action
on github itself without require creating a release.
Once we have tested it and everything is running
fine, we will remove the manual bypass of
is release step.