Automate desktop app release process (#361)
* chore: add publish app github action

Add package and publish action
for production desktop app.
This action is triggered once we have
a Release App <version> commit into
the app-stable branch. We are implementing
a git flow approach where app-stable will
be the production branch.

* chore: add dev package app action

We want to be able to test
packaging the app regularly. This
means that we need to have an action
that does that without actually going
through the process of signing the
packaged app with the production
certificates.

* chore: set desktop app version to 0.0.1

This is just so that we can initialize
the desktop app with the create-release-branch.

* chore: add github token to package app action

Add github token with write contents
permissions because electron-builder
will publish the artifacts to the existing
draft release

* chore: allow manually bypassing is release

We want to do this for testing purposes.
Testing the package app locally with act
is not possible (act is super limited.. for
example it does not support multiple OS, only linux).
For that reason we want to test the action
on github itself without require creating a release.
Once we have tested it and everything is running
fine, we will remove the manual bypass of
is release step.