fix: remove package-app workflow
Remove the generic package app
callable workflow because it is not
possible to use environments with secrets on
the calling workflow. This means that
we would need to differentiate envs
on the reusable workflow. This is something
that we want to avoid, as we do not want
unintended jobs to run with the production
env secrets.