chore: only copy .metamaskrc file if exists
Currently we need to run
the init extension on the CI
in order to have the common package
available as an extension deps.
However the .metamaskrc file does
not exist when running on the CI.
So we need to only copy the file
when it exists. Otherwise we are
displaying a warning stating that
the file was not found