fix: remove package-app workflow (#380)
* fix: remove package-app workflow

Remove the generic package app
callable workflow because it is not
possible to use environments with secrets on
the calling workflow. This means that
we would need to differentiate envs
on the reusable workflow. This is something
that we want to avoid, as we do not want
unintended jobs to run with the production
env secrets.

* chore: update package jobs

We need to take into acccount init
the submodule and initialize it properly
(after it was updated to yarn 3)

* chore: only copy .metamaskrc file if exists

Currently we need to run
the init extension on the CI
in order to have the common package
available as an extension deps.
However the .metamaskrc file does
not exist when running on the CI.
So we need to only copy the file
when it exists. Otherwise we are
displaying a warning stating that
the file was not found