chore: change release branch name to use package version
Create release branch works with
the idea of sequential releases. The
reason for that, is that it was created
with support for single releases (even if
we are working on a monorepo). However
on our specific case we do independent
workspace releases. So in order to break
from sequential releases and support multiple
ongoing parallel releases, we want the release
branch to follow the release/workspace@new_version
format.