chore: patch webextension polyfill
electron BrowserWindows do not
have the property chrome.runtime.
In fact, unless we are loading an
extension, it will not have any of
the browser extension API properties.
In order to polyfill those we are
relying on the webextension-polyfill
package. But unlike the background
process, we do not have chrome.runtime
within electron Browser window (as there
is no extension id, because we are not
loading an extension). We are patching
webextension-polyfill to bypass checking
for chrome.runtime and chrome.runtime.id