desktop popup poc (#532)
* chore: initial draft of desktop popup

* chore: use extension locale duck

* chore: use extension selectors

Show and hide window using background controller listeners.

* chore: show and hide approval window by intercepting existing popup logic

* fix: paths after structure refactor

* fix: lavamoat policies

* chore: patch webextension polyfill

electron BrowserWindows do not
have the property chrome.runtime.
In fact, unless we are loading an
extension, it will not have any of
the browser extension API properties.
In order to polyfill those we are
relying on the webextension-polyfill
package. But unlike the background
process, we do not have chrome.runtime
within electron Browser window (as there
is no extension id, because we are not
loading an extension). We are patching
webextension-polyfill to bypass checking
for chrome.runtime and chrome.runtime.id

* test: add env to disable extension notification popup

* fix: notification manager expect return window

Notification manager expects
a window object with left prop
to be returned when creating
a popup window.

* feat: close popup when no pending actions

Replicate the behaviour from
home component on the extension.
Confirmation and permission pages
route to the '/' (home) page. And
then the home page decides if the
notification popup window should be
closed or not.

* fix: prevent global.platform on approval window

There are a couple of actions
that call methods from global.platform.
Those are not available in the electron
BrowserWindow and we do not need them.
To prevent unhandled promise rejections
we are overriding those to noop

* fix: build type on desktop ui

* fix: linting

* chore: split popup-ui from desktop-ui

* fix: add root reducer to popup ui store

* chore: split styles bundle

Split styles bundle for both
desktop ui and popup ui.

* chore: update approval window when theme changes

* chore: fix styling of snap dialogs

* chore: position approval window based on extension popup

* chore: request state when showing desktop popup

* chore: shim windows.update

* chore: shim tabs query

* fix: add popup browser shim

This will allow us to shim the
browser api within the desktop
popup window without requiring
any changes to the Extension repo

* fix: lavamoat policies

* fix: styling of popup for v3 and v3 signing

* fix: linting and unit tests

* fix: e2e tests

* fix: update window handler returns

if desktop popup is not enabled
the notification-manager is still
expecting a window create/update/
remove object to be returned

* fix: symlinks and rebase

* fix: validate lavamoat ci

We need lavamoat policies when
the extension react deps are symlink
to the app deps

* fix: lint tsc

Conscious ignoring popup-ui
as it was causing issues due to
imports from the extension submodule.
And we are not using TS on the popup-ui
yet.

---------

Co-authored-by: Matthew Walsh <matthew.walsh@consensys.net>