fix: trigger extension ui when desktop popup is disabled
The disableExtensionPopup is only validating if
the env vars for desktop popup and disable extension
popup are enabled. Then on runtime, the desktop popup
can be disabled on by the user on the settings screen.
For that reason we need to check if the desktop window
handlers are registered or not. If not, it means that
the desktop popup is currently disabled on the settings
and we want to trigger the extension popup