fix: add a sleep before enable desktop popup
While testing this feature, realised
that there is a race condition between
the renderer process updating the desktop
popup enabled state (and persisting it)
and the main process starting the process of
enabling the desktop popup (which requires
emiting a restart in order to retrigger the
connectRemote for the new desktop popup renderer
process).
Solving this with a quick hack like adding this sleep
as the desktop popup feature is highly
experimental and we want to ship the first version of
desktop. Furthermore, the impact to the users should
be 0, as currently when enabling, the user needs to
re-introduce the password into the Extension UI
(in order to unlock the vault). Doing that will
take longer than the sleep that we are adding