Chore/package with desktop popup feature enabled (#601)
* chore: add desktop popup vars to package actions

Add desktop popup enable env vars
to github package actions. The actions
will get the values from the respective
github environment variables.

* fix: trigger extension ui when desktop popup is disabled

The disableExtensionPopup is only validating if
the env vars for desktop popup and disable extension
popup are enabled. Then on runtime, the desktop popup
can be disabled on by the user on the settings screen.
For that reason we need to check if the desktop window
handlers are registered or not. If not, it means that
the desktop popup is currently disabled on the settings
and we want to trigger the extension popup

* fix: desktop build type

* fix: add a sleep before enable desktop popup

While testing this feature, realised
that there is a race condition between
the renderer process updating the desktop
popup enabled state (and persisting it)
and the main process starting the process of
enabling the desktop popup (which requires
emiting a restart in order to retrigger the
connectRemote for the new desktop popup renderer
process).
Solving this with a quick hack like adding this sleep
as the desktop popup feature is highly
experimental and we want to ship the first version of
desktop. Furthermore, the impact to the users should
be 0, as currently when enabling, the user needs to
re-introduce the password into the Extension UI
(in order to unlock the vault). Doing that will
take longer than the sleep that we are adding

* fix: exclude vuln GHSA-p8p7-x288-28g6

This is cause by @metamask/network-controller
importing an old web3-provider-engine maintained
by metamask. That package is still using
the long deprecated request package.
By passing the issue here, until it is fixed within
that package.