fix: add babel plugin to remove explicit ts import
Currently the extension has explicit ts
import files (example - import * 'test.ts').
This is not a good pattern for node, as the
ts build will output js files. And thus during
the runtime the node process breaks as it cannot
find the imported file. On the Extension it does not
have any issues because of the bundling that we have
in place. It creates a reference to the path and as so
it can have ts and js files beig explicitly imported,
but in reality it is only running js.
The fix that we are doing here, is just to
remove the .ts file extension from any import (as the
ts file itself is already transpiled into js and
has the js file extension)