fix: get NFT image from data base64
The NFTController is always fetching the
NFT image, even if the image is a data base64
encoded. Fetching anything other than http/https
is supported in the browser but not by node-fetch
(which uses the underlying http/https node native modules).
To fix this we will use the native node fetch
(available by default since node 18) to perform
any fetch actions that are not http/https.