fix: runtime getUrl
This method is not async in the browser.
Due to that on the Extension we are calling
it synchronously to go back to the home page
(from the phishing blocked script). When on Desktop
we proxy this requests to the actual browser, meaning
that the request will be async. This was causing
the phishing detection to not return to the extension
home page when clicking on the back to safety button.
Also, it is worth mentioning that the
protocol for chrome and firefox extensions are different.
So added a method to update the browser info
accordingly and route to the correct url.